requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pymongo[srv]==4.8.0
langdetect==1.0.9
transformers==4.43.3
torch>=2.2.0
google-generativeai==0.7.2
python-dotenv==1.0.1
gunicorn==21.2.0
